// src/shared/types/notification.ts
export interface RemoteMessage {
  /**
   * The collapse key a message was sent with. Used to override existing messages with the same
   * key.
   */
  collapseKey?: string;

  /**
   * A unique ID assigned to every message.
   *
   * If not provided, a random unique ID is generated.
   */
  messageId?: string;

  /**
   * The message type of the message.
   */
  messageType?: string;
  /**
   * The topic name or message identifier.
   */
  from?: string;
  /**
   * The address for the message.
   */
  to?: string;

  /**
   * The time to live for the message in seconds.
   *
   * Defaults to 3600.
   */
  ttl?: number;

  /**
   * The time the message was sent, in milliseconds since the start of unix epoch
   */
  sentTime?: number;

  /**
   * Any additional data sent with the message.
   */
  data?: {[key: string]: string | object};

  /**
   * Additional NotificationPayload data sent with the message
   */
  notification?: NotificationPayload;

  /**
   * Whether the iOS APNs message was configured as a background update notification.
   *
   * @platform ios iOS
   */
  contentAvailable?: boolean;

  /**
   * Whether the iOS APNs `mutable-content` property on the message was set
   * allowing the app to modify the notification via app extensions.
   *
   * @platform ios iOS
   */
  mutableContent?: boolean;

  /**
   * The iOS category this notification is assigned to.
   *
   * @platform ios iOS
   */
  category?: string;

  /**
     * An iOS app specific identifier used for notification grouping.
    threadId?: string;
    */
  threadId?: string;

  /**
   * Options for features provided by the FCM SDK for Web.
   */
  // fcmOptions: FcmOptions;
}

export interface FcmOptions {
  /**
   * The link to open when the user clicks on the notification.
   */
  link?: string;

  /**
   * The label associated with the message's analytics data.
   */
  analyticsLabel?: string;
}

/**
 * NotificationPayload is an alias for Notification. This is to keep it the same as
 * Firebase Web JS SDK v9 and to make it backwards compatible.
 */
type NotificationPayload = Notification;

export interface Notification {
  /**
   * The native localization key for the notification title.
   */
  titleLocKey?: string;

  /**
   * Any arguments that should be formatted into the resource specified by titleLocKey.
   */
  titleLocArgs?: string[];

  /**
   * The notification body content.
   */
  body?: string;

  /**
   * Web only. The URL to use for the notification's icon. If you don't send this key in the request,
   * FCM displays the launcher icon specified in your app manifest.
   */
  icon?: string;

  /**
   * Web only. The URL of an image that is downloaded on the device and displayed in the notification.
   */
  image?: string;

  /**
   * Web only. The notification's title.
   */
  title?: string;

  /**
   * The native localization key for the notification body content.
   */
  bodyLocKey?: string;

  /**
   * Any arguments that should be formatted into the resource specified by bodyLocKey.
   */
  bodyLocArgs?: string[];

  ios?: {
    /**
     * The notification's subtitle.
     */
    subtitle?: string;

    /**
     * The native localization key for the notification's subtitle.
     */
    subtitleLocKey?: string;

    /**
     * Any arguments that should be formatted into the resource specified by subtitleLocKey.
     */
    subtitleLocArgs?: string[];

    /**
     * The value of the badge on the home screen app icon.
     * If not specified, the badge is not changed.
     * If set to 0, the badge has been removed.
     */
    badge?: string;

    /**
     * The sound played when the notification was delivered on the device (if permissions permit).
     */
    sound?: string | NotificationIOSCriticalSound;
  };

  /**
   * Additional Android specific properties set on the notification.
   */
  android?: {
    /**
     * The sound played when the notification was delivered on the device (channel settings permitted).
     *
     * Set as "default" if the default device notification sound was used.
     */
    sound?: string;

    /**
     * The channel ID set on the notification. If not set, the notification uses the default
     * "Miscellaneous" channel set by FCM.
     */
    channelId?: string;

    /**
     * The custom color used to tint the notification content.
     */
    color?: string;

    /**
     * The custom small icon used to display on the notification. If not set, uses the default
     * application icon defined in the AndroidManifest file.
     */
    smallIcon?: string;

    /**
     * The custom image was provided and displayed in the notification body.
     */
    imageUrl?: string;

    /**
     * Deep link URL provided to the notification.
     */
    link?: string;

    /**
     * The current unread notification count for this application, managed by the device.
     */
    count?: number;

    /**
     * Name of the click action set on the notification.
     */
    clickAction?: string;

    /**
     * The notification priority.
     *
     * Note; on devices which have channel support (Android 8.0 (API level 26) +),
     * this value will be ignored. Instead, the channel "importance" level is used.
     */
    priority?:
      | NotificationAndroidPriority.PRIORITY_MIN
      | NotificationAndroidPriority.PRIORITY_LOW
      | NotificationAndroidPriority.PRIORITY_DEFAULT
      | NotificationAndroidPriority.PRIORITY_HIGH
      | NotificationAndroidPriority.PRIORITY_MAX;

    /**
     * Ticker text set on the notification.
     *
     * Ticker text is used for devices with accessibility enabled (e.g. to read out the message).
     */
    ticker?: string;

    /**
     * The visibility of a notification. This value determines how the notification is shown on the users
     * devices (e.g. on the lock-screen).
     */
    visibility?:
      | NotificationAndroidVisibility.VISIBILITY_SECRET
      | NotificationAndroidVisibility.VISIBILITY_PRIVATE
      | NotificationAndroidVisibility.VISIBILITY_PUBLIC;
  };
}